backend/pyproject.toml
[project]
name = "python-fastapi"
version = "0.1.0"
description = "FastAPI for the 'Academic Neighbour' project"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[all]>=0.119.1",
    "psycopg2>=2.9.11",
    "pwdlib[argon2]>=0.3.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "ruff>=0.14.3",
    "sqlmodel>=0.0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "ruff>=0.3.0",
    "black>=24.0.0",
    "isort>=5.13.0",
    "mypy>=1.8.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --strict-markers --tb=short"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "SIM", "UP"]

[tool.coverage.run]
source = ["."]
omit = ["tests/*", "venv/*", ".venv/*"]
